refactor(resolver): resolver를 rest용, websocket용으로 2개를 만들어 사용
