+++ src/main/java/com/example/knu_connect/global/config/WebConfig.java
@@ -1,6 +1,6 @@
 package com.example.knu_connect.global.config;
 
-import com.example.knu_connect.global.resolver.AuthUserArgumentResolver;
+import com.example.knu_connect.global.resolver.RestAuthUserArgumentResolver;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
@@ -12,10 +12,10 @@ import java.util.List;
 @RequiredArgsConstructor
 public class WebConfig implements WebMvcConfigurer {
 
-    private final AuthUserArgumentResolver authUserArgumentResolver;
+    private final RestAuthUserArgumentResolver restAuthUserArgumentResolver;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(authUserArgumentResolver);
+        resolvers.add(restAuthUserArgumentResolver);
     }
 }
