+++ src/main/java/com/example/knu_connect/global/resolver/WebSocketAuthUserArgumentResolver.java
@@ -1,0 +1,43 @@
+package com.example.knu_connect.global.resolver;
+
+import com.example.knu_connect.domain.user.entity.User;
+import com.example.knu_connect.global.annotation.AuthUser;
+import com.example.knu_connect.global.auth.jwt.CustomUserDetails;
+import com.example.knu_connect.global.exception.common.BusinessException;
+import com.example.knu_connect.global.exception.common.ErrorCode;
+import org.springframework.core.MethodParameter;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.handler.invocation.HandlerMethodArgumentResolver;
+import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
+import org.springframework.messaging.support.MessageHeaderAccessor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.stereotype.Component;
+
+@Component
+public class WebSocketAuthUserArgumentResolver implements HandlerMethodArgumentResolver {
+
+    @Override
+    public boolean supportsParameter(MethodParameter parameter) {
+        return parameter.hasParameterAnnotation(AuthUser.class)
+                && User.class.isAssignableFrom(parameter.getParameterType());
+    }
+
+    @Override
+    public Object resolveArgument(MethodParameter parameter, Message<?> message) throws Exception {
+        StompHeaderAccessor accessor = MessageHeaderAccessor.getAccessor(message, StompHeaderAccessor.class);
+
+        if (accessor != null && accessor.getUser() != null) {
+            Object user = accessor.getUser();
+
+            if (user instanceof UsernamePasswordAuthenticationToken auth) {
+                Object principal = auth.getPrincipal();
+
+                if (principal instanceof CustomUserDetails customUserDetails) {
+                    return customUserDetails.getUser();
+                }
+            }
+        }
+
+        throw new BusinessException(ErrorCode.UNAUTHORIZED_ACCESS);
+    }
+}
